Update patch set 3
Patch Set 3: Code-Review+1

I don't think that's what he was talking about.  I believe that case will be handled on the next boot - MediaScanner will see that the media is unmounted and clear it from the DB.

Problem is without this there is a race condition between shutdown and that eject event - which could cause the device to clear the media DB when it shouldn't be cleared.

Patch-set: 3